Move split proposal word tests to correct class.
Word proposal contained tests for proposal and submitted proposal (word). Thus
the test-cases are moved to the corresponding class.
